requirements.txt
openpyxl==3.0.9
pandas==1.3.5
Pillow==8.4.0
requests==2.26.0
requests-toolbelt==0.9.1
urllib3==1.26.7
python-dotenv==0.19.2
